move 1.0.0 to a version directory
