+++ crates/components-runtime/src/data_collection/versions/mod.rs
@@ -1,5 +1,7 @@
 use serde::Deserialize;
 
+pub mod v1_0_0;
+
 #[derive(Debug, Clone, Deserialize, Default, PartialEq)]
 pub enum DataCollectionWitVersion {
     #[default]
